WordPress has been updated with latest version 4.9.1

+++ wp-admin/includes/class-wp-comments-list-table.php
@@ -29,7 +29,6 @@ class WP_Comments_List_Table extends WP_List_Table {
 	 * Constructor.
 	 *
 	 * @since 3.1.0
-	 * @access public
 	 *
 	 * @see WP_List_Table::__construct() for more information on default arguments.
 	 *
@@ -94,7 +93,7 @@ class WP_Comments_List_Table extends WP_List_Table {
 
 		$comments_per_page = $this->get_per_page( $comment_status );
 
-		$doing_ajax = defined( 'DOING_AJAX' ) && DOING_AJAX;
+		$doing_ajax = wp_doing_ajax();
 
 		if ( isset( $_REQUEST['number'] ) ) {
 			$number = (int) $_REQUEST['number'];
@@ -166,7 +165,7 @@ class WP_Comments_List_Table extends WP_List_Table {
 	public function get_per_page( $comment_status = 'all' ) {
 		$comments_per_page = $this->get_items_per_page( 'edit_comments_per_page' );
 		/**
-		 * Filter the number of comments listed per page in the comments list table.
+		 * Filters the number of comments listed per page in the comments list table.
 		 *
 		 * @since 2.6.0
 		 *
@@ -201,8 +200,7 @@ class WP_Comments_List_Table extends WP_List_Table {
 
 		$status_links = array();
 		$num_comments = ( $post_id ) ? wp_count_comments( $post_id ) : wp_count_comments();
-		//, number_format_i18n($num_comments->moderated) ), "<span class='comment-count'>" . number_format_i18n($num_comments->moderated) . "</span>"),
-		//, number_format_i18n($num_comments->spam) ), "<span class='spam-comment-count'>" . number_format_i18n($num_comments->spam) . "</span>")
+
 		$stati = array(
 			/* translators: %s: all comments count */
 			'all' => _nx_noop(
@@ -248,7 +246,11 @@ class WP_Comments_List_Table extends WP_List_Table {
 			$link = add_query_arg( 'comment_type', $comment_type, $link );
 
 		foreach ( $stati as $status => $label ) {
-			$class = ( $status === $comment_status ) ? ' class="current"' : '';
+			$current_link_attributes = '';
+
+			if ( $status === $comment_status ) {
+				$current_link_attributes = ' class="current" aria-current="page"';
+			}
 
 			if ( !isset( $num_comments->$status ) )
 				$num_comments->$status = 10;
@@ -260,7 +262,7 @@ class WP_Comments_List_Table extends WP_List_Table {
 			if ( !empty( $_REQUEST['s'] ) )
 				$link = add_query_arg( 's', esc_attr( wp_unslash( $_REQUEST['s'] ) ), $link );
 			*/
-			$status_links[ $status ] = "<a href='$link'$class>" . sprintf(
+			$status_links[ $status ] = "<a href='$link'$current_link_attributes>" . sprintf(
 				translate_nooped_plural( $label, $num_comments->$status ),
 				sprintf( '<span class="%s-count">%s</span>',
 					( 'moderated' === $status ) ? 'pending' : $status,
@@ -270,7 +272,7 @@ class WP_Comments_List_Table extends WP_List_Table {
 		}
 
 		/**
-		 * Filter the comment status links.
+		 * Filters the comment status links.
 		 *
 		 * @since 2.5.0
 		 *
@@ -320,6 +322,11 @@ class WP_Comments_List_Table extends WP_List_Table {
 	 */
 	protected function extra_tablenav( $which ) {
 		global $comment_status, $comment_type;
+		static $has_items;
+
+		if ( ! isset( $has_items ) ) {
+			$has_items = $this->has_items();
+		}
 ?>
 		<div class="alignleft actions">
 <?php
@@ -330,7 +337,7 @@ class WP_Comments_List_Table extends WP_List_Table {
 				<option value=""><?php _e( 'All comment types' ); ?></option>
 <?php
 				/**
-				 * Filter the comment types dropdown menu.
+				 * Filters the comment types dropdown menu.
 				 *
 				 * @since 2.7.0
 				 *
@@ -352,10 +359,10 @@ class WP_Comments_List_Table extends WP_List_Table {
 			 * @since 3.5.0
 			 */
 			do_action( 'restrict_manage_comments' );
-			submit_button( __( 'Filter' ), 'button', 'filter_action', false, array( 'id' => 'post-query-submit' ) );
+			submit_button( __( 'Filter' ), '', 'filter_action', false, array( 'id' => 'post-query-submit' ) );
 		}
 
-		if ( ( 'spam' === $comment_status || 'trash' === $comment_status ) && current_user_can( 'moderate_comments' ) ) {
+		if ( ( 'spam' === $comment_status || 'trash' === $comment_status ) && current_user_can( 'moderate_comments' ) && $has_items ) {
 			wp_nonce_field( 'bulk-destroy', '_destroy_nonce' );
 			$title = ( 'spam' === $comment_status ) ? esc_attr__( 'Empty Spam' ) : esc_attr__( 'Empty Trash' );
 			submit_button( $title, 'apply', 'delete_all', false );
@@ -424,7 +431,6 @@ class WP_Comments_List_Table extends WP_List_Table {
 	 * Get the name of the default primary column.
 	 *
 	 * @since 4.3.0
-	 * @access protected
 	 *
 	 * @return string Name of the default primary column, in this case, 'comment'.
 	 */
@@ -433,7 +439,6 @@ class WP_Comments_List_Table extends WP_List_Table {
 	}
 
 	/**
-	 * @access public
 	 */
 	public function display() {
 		wp_nonce_field( "fetch-list-" . get_class( $this ), '_ajax_fetch_list_nonce' );
@@ -506,7 +511,6 @@ class WP_Comments_List_Table extends WP_List_Table {
  	 * Generate and display row actions links.
  	 *
  	 * @since 4.3.0
- 	 * @access protected
  	 *
  	 * @global string $comment_status Status for the current listed comments.
  	 *
@@ -601,7 +605,7 @@ class WP_Comments_List_Table extends WP_List_Table {
 			( ( ( 'approve' === $action || 'unapprove' === $action ) && 2 === $i ) || 1 === $i ) ? $sep = '' : $sep = ' | ';
 
 			// Reply and quickedit need a hide-if-no-js span when not added with ajax
-			if ( ( 'reply' === $action || 'quickedit' === $action ) && ! defined('DOING_AJAX') )
+			if ( ( 'reply' === $action || 'quickedit' === $action ) && ! wp_doing_ajax() )
 				$action .= ' hide-if-no-js';
 			elseif ( ( $action === 'untrash' && $the_comment_status === 'trash' ) || ( $action === 'unspam' && $the_comment_status === 'spam' ) ) {
 				if ( '1' == get_comment_meta( $comment->comment_ID, '_wp_trash_meta_status', true ) )
@@ -691,7 +695,7 @@ class WP_Comments_List_Table extends WP_List_Table {
 
 		if ( $this->user_can ) {
 			if ( ! empty( $comment->comment_author_email ) ) {
-				/* This filter is documented in wp-includes/comment-template.php */
+				/** This filter is documented in wp-includes/comment-template.php */
 				$email = apply_filters( 'comment_email', $comment->comment_author_email, $comment );
 
 				if ( ! empty( $email ) && '@' !== $email ) {
@@ -711,14 +715,13 @@ class WP_Comments_List_Table extends WP_List_Table {
 	}
 
 	/**
-	 * @access public
 	 *
 	 * @param WP_Comment $comment The comment object.
 	 */
 	public function column_date( $comment ) {
 		/* translators: 1: comment date, 2: comment time */
 		$submitted = sprintf( __( '%1$s at %2$s' ),
-			/* translators: comment date format. See http://php.net/date */
+			/* translators: comment date format. See https://secure.php.net/date */
 			get_comment_date( __( 'Y/m/d' ), $comment ),
 			get_comment_date( __( 'g:i a' ), $comment )
 		);
@@ -737,7 +740,6 @@ class WP_Comments_List_Table extends WP_List_Table {
 	}
 
 	/**
-	 * @access public
 	 *
 	 * @param WP_Comment $comment The comment object.
 	 */
